+++ OsmAnd/src/net/osmand/plus/wikivoyage/explore/travelcards/TravelNeededMapsCard.java
@@ -154,8 +154,8 @@ public class TravelNeededMapsCard extends BaseTravelCard {
 			ProgressBar progressBar = view.findViewById(R.id.progress_bar);
 			progressBar.setVisibility(downloading ? View.VISIBLE : View.GONE);
 			if (currentDownloading) {
-				float progress = downloadThread.getCurrentDownloadProgress();
-				progressBar.setProgress(progress < 0 ? 0 : (int) progress);
+				int progress = (int) downloadThread.getCurrentDownloadProgress();
+				progressBar.setProgress(progress < 0 ? 0 : progress);
 			} else {
 				progressBar.setProgress(0);
 			}
